move sections to sub namespace
