+++ dotnet/packages/Microsoft.TeamsAI/Microsoft.TeamsAI.Tests/AITests/PromptsTests/SectionsTests/TemplateSectionTests.cs
@@ -1,10 +1,10 @@
 ﻿using Azure.AI.OpenAI;
 using Microsoft.Bot.Builder;
-using Microsoft.Teams.AI.AI.Prompts;
+using Microsoft.Teams.AI.AI.Prompts.Sections;
 using Microsoft.Teams.AI.AI.Tokenizers;
 using Moq;
 
-namespace Microsoft.Teams.AI.Tests.AITests.PromptsTests
+namespace Microsoft.Teams.AI.Tests.AITests.PromptsTests.SectionsTests
 {
     public class TemplateSectionTests
     {
