+++ dotnet/packages/Microsoft.TeamsAI/Microsoft.TeamsAI/AI/Prompts/Sections/AssistantMessageSection.cs
@@ -1,19 +1,19 @@
 ﻿using Azure.AI.OpenAI;
 
-namespace Microsoft.Teams.AI.AI.Prompts
+namespace Microsoft.Teams.AI.AI.Prompts.Sections
 {
     /// <summary>
-    /// A user message section.
+    /// An assistant message section.
     /// </summary>
-    public class UserMessageSection : TemplateSection
+    public class AssistantMessageSection : TemplateSection
     {
         /// <summary>
-        /// Creates an instance of `UserMessageSection`
+        /// Creates an instance of `AssistantMessageSection`
         /// </summary>
         /// <param name="template">Template to use for this section.</param>
         /// <param name="tokens">Sizing strategy for this section. Defaults to `auto`.</param>
         /// <param name="prefix">Prefix to use for user messages when rendering as text. Defaults to `user: `.</param>
-        public UserMessageSection(string template, int tokens = -1, string prefix = "user: ") : base(template, ChatRole.User, tokens, true, "\n", prefix)
+        public AssistantMessageSection(string template, int tokens = -1, string prefix = "assistant: ") : base(template, ChatRole.Assistant, tokens, true, "\n", prefix)
         {
         }
     }
