+++ dotnet/packages/Microsoft.TeamsAI/Microsoft.TeamsAI/AI/Prompts/Sections/UserMessageSection.cs
@@ -1,19 +1,19 @@
 ﻿using Azure.AI.OpenAI;
 
-namespace Microsoft.Teams.AI.AI.Prompts
+namespace Microsoft.Teams.AI.AI.Prompts.Sections
 {
     /// <summary>
-    /// An assistant message section.
+    /// A user message section.
     /// </summary>
-    public class AssistantMessageSection : TemplateSection
+    public class UserMessageSection : TemplateSection
     {
         /// <summary>
-        /// Creates an instance of `AssistantMessageSection`
+        /// Creates an instance of `UserMessageSection`
         /// </summary>
         /// <param name="template">Template to use for this section.</param>
         /// <param name="tokens">Sizing strategy for this section. Defaults to `auto`.</param>
         /// <param name="prefix">Prefix to use for user messages when rendering as text. Defaults to `user: `.</param>
-        public AssistantMessageSection(string template, int tokens = -1, string prefix = "assistant: ") : base(template, ChatRole.Assistant, tokens, true, "\n", prefix)
+        public UserMessageSection(string template, int tokens = -1, string prefix = "user: ") : base(template, ChatRole.User, tokens, true, "\n", prefix)
         {
         }
     }
